+++ src/main/java/org/bytedeco/javacpp/tools/Parser.java
@@ -377,7 +377,7 @@ public class Parser {
             } else if (token.match(Token.VIRTUAL)) {
                 type.virtual = true;
             } else if (token.match(Token.ENUM, Token.EXPLICIT, Token.EXTERN, Token.INLINE, Token.CLASS, Token.INTERFACE,
-                                   Token.STRUCT, Token.UNION, Token.TYPEDEF, Token.TYPENAME, Token.USING)) {
+                                   Token.__INTERFACE, Token.STRUCT, Token.UNION, Token.TYPEDEF, Token.TYPENAME, Token.USING)) {
                 continue;
             } else if (token.match((Object[])InfoMap.simpleTypes) && (type.cppName.length() == 0 || type.simple)) {
                 type.cppName += token.value + " ";
@@ -1816,9 +1816,9 @@ public class Parser {
         boolean foundGroup = false, friend = false;
         Context ctx = new Context(context);
         for (Token token = tokens.get(); !token.match(Token.EOF); token = tokens.next()) {
-            if (token.match(Token.CLASS, Token.INTERFACE, Token.STRUCT, Token.UNION)) {
+            if (token.match(Token.CLASS, Token.INTERFACE, Token.__INTERFACE, Token.STRUCT, Token.UNION)) {
                 foundGroup = true;
-                ctx.inaccessible = token.match(Token.CLASS, Token.INTERFACE);
+                ctx.inaccessible = token.match(Token.CLASS);
                 break;
             } else if (token.match(Token.FRIEND)) {
                 friend = true;