+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Fix `Parser` support for the `interface` keyword of the Microsoft C/C++ Compiler
  * Let `Parser` pick up names from `Info.pointerTypes` in the case of function pointers as well
  * Add new "org.bytedeco.javacpp.nopointergc" system property to prevent `Pointer` from registering deallocators with the garbage collector
  * Add `@Properties(names=...)` value to specify a list of default platform names that can be inherited by other classes